fix: fix #250 format numberic result
