+++ src/contants.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const AGGREGATOR_MAP = {
+  count: 'count',
+  average: 'avg',
+  minimum: 'min',
+  maximum: 'max',
+  sum: 'sum'
+};
+
+module.exports = {
+  AGGREGATOR_MAP
+};
